Fix enum underlying types: resolve _MTL_OPTIONS typedefs, use ulong for NS::UInteger, use ValueExpression for large values

+++ Metal.NET/Metal/MTL4AlphaToCoverageState.cs
@@ -1,6 +1,6 @@
 ﻿namespace Metal.NET;
 
-public enum MTL4AlphaToCoverageState : int
+public enum MTL4AlphaToCoverageState : long
 {
     AlphaToCoverageStateDisabled = 0,
