+++ Metal.NET/Metal/MTLCommandEncoder.cs
@@ -30,7 +30,7 @@ public class MTLCommandEncoder : IDisposable
 
     public void BarrierAfterQueueStages(MTLStages afterQueueStages, MTLStages beforeStages)
     {
-        ObjectiveCRuntime.MsgSend(NativePtr, MTLCommandEncoderSelector.BarrierAfterQueueStagesBeforeStages, (uint)afterQueueStages, (uint)beforeStages);
+        ObjectiveCRuntime.MsgSend(NativePtr, MTLCommandEncoderSelector.BarrierAfterQueueStagesBeforeStages, (nuint)afterQueueStages, (nuint)beforeStages);
     }
 
     public void EndEncoding()
