+++ Metal.NET/Metal/MTLFunctionLogType.cs
@@ -1,6 +1,6 @@
 ﻿namespace Metal.NET;
 
-public enum MTLFunctionLogType : uint
+public enum MTLFunctionLogType : ulong
 {
     FunctionLogTypeValidation = 0
 }
